+++ src/IPFees.Calculator/Evaluator/Node.cs
@@ -3,6 +3,6 @@
     // Node - abstract class representing one node in the expression 
     public abstract class Node
     {
-        public abstract double Eval(IContext ctx);
+        public abstract decimal Eval(IContext ctx);
     }
 }
